requirements.txt
async-generator==1.10
attrs==21.4.0
certifi==2022.6.15
cffi==1.15.0
click==8.1.3
colorama==0.4.5
cryptography==37.0.2
cycler==0.11.0
Flask==2.1.2
fonttools==4.33.3
h11==0.13.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.1
kiwisolver==1.4.3
MarkupSafe==2.1.1
matplotlib==3.5.2
MouseInfo==0.1.3
numpy==1.23.0
opencv-python==4.6.0.66
outcome==1.2.0
packaging==21.3
pandas==1.4.3
Pillow==9.1.1
PyAutoGUI==0.9.53
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.9
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.28
PySocks==1.7.1
python-dateutil==2.8.2
pytweening==1.0.4
pytz==2022.1
PyYAML==6.0
selenium==4.3.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.21.0
trio-websocket==0.9.2
urllib3==1.26.9
Werkzeug==2.1.2
wordcloud==1.8.1
wsproto==1.1.0
zipp==3.8.0
